Version 1.0: This program inserts a pseudo random value into a pseudo random location within each of the 9 'boxes' in the sudoku grid as long as they will yield a valid solution. Using this created puzzle the program proceeds to find a solution.
This code implements the backtracking algorithm to solve any given sudoku puzzle.
This implementation is essentially a binary tree. It begins by finding the first empty cell indicated by a '0'. If no empty values are found then it is declared solved. A for loop finds the lowest valid number for that cell. A number is 'valid' at a particular cell if that number does not exist within the same 1x9 row or 1x9 column as that index, as well as within the same 3x3 square that it occupies. This value is inserted into the grid and the solution continues assuming this value was correct. If at some point the solution runs out of potential values to use at a cell it will revert to the previous cell and try a different value at that index. It will continue checking each path in the tree until a solution is found. If there are multiple unique solutions it will only find one, with a tendency to find solutions with lower values in the upper left corner and higher values in the lower right corner due to the nature of the algorithm.